Renaming of files to imu_sensor.
Add of imu_sensor_with_sensitivity
